refactor: cleanup adjust threshold handler

+++ apps/api/Modules/Inventory/Inventory/Products/Features/AdjustProductStock/AdjustProductStockHandler.cs
@@ -1,6 +1,6 @@
 ﻿namespace Inventory.Products.Features.AdjustProductStock;
 
-public record AdjustProductStockCommand(Guid Id, decimal Delta)
+public record AdjustProductStockCommand(Guid Id, int Delta)
     : ICommand<AdjustProductStockResult>;
 
 public record AdjustProductStockResult(Guid Id);
